+++ grid2op/Environment/__init__.py
@@ -1,9 +1,13 @@
 __all__ = [
     "BaseEnv",
     "Environment",
-    "MultiEnvironment"
+    "BaseMultiProcessEnvironment",
+    "SingleEnvMultiProcess",
+    "MultiEnvMultiProcess"
 ]
 
 from grid2op.Environment.BaseEnv import BaseEnv
 from grid2op.Environment.Environment import Environment
-from grid2op.Environment.MultiEnv import MultiEnvironment
+from grid2op.Environment.BaseMultiProcessEnv import BaseMultiProcessEnvironment
+from grid2op.Environment.SingleEnvMultiProcess import SingleEnvMultiProcess
+from grid2op.Environment.MultiEnvMultiProcess import MultiEnvMultiProcess