+++ grid2op/Reward/CloseToOverflowReward.py
@@ -15,11 +15,11 @@ class CloseToOverflowReward(BaseReward):
     Returns max reward when there is no overflow, min reward if more than one line is close to overflow
     and the mean between max and min reward if one line is close to overflow
     """
-    def __init__(self):
+    def __init__(self, max_lines=5):
         BaseReward.__init__(self)
         self.reward_min = dt_float(0.0)
         self.reward_max = dt_float(1.0)
-        self.max_overflowed = dt_float(5.0)
+        self.max_overflowed = dt_float(max_lines)
 
     def initialize(self, env):
         pass
@@ -35,7 +35,7 @@ class CloseToOverflowReward(BaseReward):
         close_to_overflow = dt_float(0.0)
         for ratio, limit in zip(lineflow_ratio, thermal_limits):
             # Seperate big line and small line
-            if (limit < 400.00 and ratio > 0.90) or ratio >= 0.95:
+            if (limit < 400.00 and ratio >= 0.95) or ratio >= 0.975:
                 close_to_overflow += dt_float(1.0)
 
         close_to_overflow = np.clip(close_to_overflow,